+++ src/main/java/com/thekey/stylekeyserver/base/BaseTimeConfig.java
@@ -1,0 +1,9 @@
+package com.thekey.stylekeyserver.base;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class BaseTimeConfig {
+}
