+++ src/main/java/com/thekey/stylekeyserver/common/properties/PropertiesConfig.java
@@ -1,0 +1,12 @@
+package com.thekey.stylekeyserver.common.properties;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableConfigurationProperties({
+        CorsProperties.class,
+        AppProperties.class
+})
+public class PropertiesConfig {
+}
