+++ src/main/java/com/thekey/stylekeyserver/common/s3/config/SchedulingConfig.java
@@ -1,0 +1,9 @@
+package com.thekey.stylekeyserver.common.s3.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@Configuration
+@EnableScheduling
+public class SchedulingConfig {
+}
